Abstracted Team Generator away
