+++ GameExecution.py
@@ -1,17 +1,7 @@
-from GameSim.Player import Player
-from GameSim.Goalie import Goalie
-from GameSim.Team import Team
-from GameSim.TeamNameGenerator import TeamNameGenerator
+from GameSim.Generators.TeamGenerator import TeamGenerator
 
 if __name__ == '__main__':
     for i in range(5):
-        t = Team()
-        TeamNameGenerator.populate_team_information(t)
-        for i in range(22):
-            p = Player()
-            t.players.append(p)
-        for i in range(3):
-            g = Goalie()
-            t.goalies.append(g)
+        t = TeamGenerator.generate_random_team()
         t.print_team()
         print()
