+++ Roadmap.txt
@@ -31,6 +31,8 @@ ADD Descriptive commentators
 
 ADD Endurance and rotational Line-changes
 
+Change Player Generation to use a generator
+
 ADD modified stats generation based on position
 
 ADD positional modifiers for situation and general
